[REF] agreement_legal: Replace repeated dynamic field code with mixin
