+++ agreement_legal/models/agreement.py
@@ -9,7 +9,8 @@ from odoo import _, api, fields, models
 
 
 class Agreement(models.Model):
-    _inherit = "agreement"
+    _name = "agreement"
+    _inherit = ["agreement", "agreement.dynamic.content.mixin"]
 
     name = fields.Char(string="Title", required=True)
     version = fields.Integer(
@@ -226,34 +227,6 @@ class Agreement(models.Model):
     )
     signed_contract_filename = fields.Char(string="Filename")
     signed_contract = fields.Binary(string="Signed Document", tracking=True)
-    field_id = fields.Many2one(
-        "ir.model.fields",
-        string="Field",
-        help="""You can select a target field from the related document model.
-        If it is a relationship field you will be able to select a target field
-        at the destination of the relationship.""",
-    )
-    sub_object_id = fields.Many2one(
-        "ir.model",
-        string="Sub-model",
-        help="""When a relationship field is selected as first field, this
-         field shows the document model the relationship goes to.""",
-    )
-    sub_model_object_field_id = fields.Many2one(
-        "ir.model.fields",
-        string="Sub-field",
-        help="""When a relationship field is selected as first field, this
-         field lets you select the target field within the destination document
-          model (sub-model).""",
-    )
-    default_value = fields.Char(
-        help="Optional value to use if the target field is empty."
-    )
-    copyvalue = fields.Char(
-        string="Placeholder Expression",
-        help="""Final placeholder expression, to be copy-pasted in the desired
-         template field.""",
-    )
     template_id = fields.Many2one(
         "agreement", string="Template", domain=[("is_template", "=", True)]
     )
@@ -321,26 +294,6 @@ class Agreement(models.Model):
                 agreement.special_terms, "agreement", [agreement.id]
             )[agreement.id]
             agreement.dynamic_special_terms = special_terms
-
-    @api.onchange("field_id", "sub_model_object_field_id", "default_value")
-    def onchange_copyvalue(self):
-        self.sub_object_id = False
-        self.copyvalue = False
-        if self.field_id and not self.field_id.relation:
-            self.copyvalue = "{{{{object.{} or {}}}}}".format(
-                self.field_id.name, self.default_value or "''"
-            )
-            self.sub_model_object_field_id = False
-        if self.field_id and self.field_id.relation:
-            self.sub_object_id = self.env["ir.model"].search(
-                [("model", "=", self.field_id.relation)]
-            )[0]
-        if self.sub_model_object_field_id:
-            self.copyvalue = "{{{{object.{}.{} or {}}}}}".format(
-                self.field_id.name,
-                self.sub_model_object_field_id.name,
-                self.default_value or "''",
-            )
 
     # Used for Kanban grouped_by view
     @api.model
