+++ agreement_legal/models/agreement_appendix.py
@@ -1,11 +1,12 @@
 # Copyright (C) 2018 - TODAY, Pavlov Media
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from odoo import api, fields, models
+from odoo import fields, models
 
 
 class AgreementAppendix(models.Model):
     _name = "agreement.appendix"
+    _inherit = ["agreement.dynamic.content.mixin"]
     _description = "Agreement Appendices"
     _order = "sequence"
 
@@ -24,56 +25,6 @@ class AgreementAppendix(models.Model):
         help="If unchecked, it will allow you to hide this appendix without "
         "removing it.",
     )
-
-    # Dynamic field editor
-    field_id = fields.Many2one(
-        "ir.model.fields",
-        string="Field",
-        help="""Select target field from the related document model. If it is a
-         relationship field you will be able to select a target field at the
-         destination of the relationship.""",
-    )
-    sub_object_id = fields.Many2one(
-        "ir.model",
-        string="Sub-model",
-        help="""When a relationship field is selected as first field, this
-         field shows the document model the relationship goes to.""",
-    )
-    sub_model_object_field_id = fields.Many2one(
-        "ir.model.fields",
-        string="Sub-field",
-        help="""When a relationship field is selected as first field, this
-         field lets you select the target field within the destination document
-          model (sub-model).""",
-    )
-    default_value = fields.Char(
-        help="Optional value to use if the target field is empty."
-    )
-    copyvalue = fields.Char(
-        string="Placeholder Expression",
-        help="""Final placeholder expression, to be copy-pasted in the desired
-         template field.""",
-    )
-
-    @api.onchange("field_id", "sub_model_object_field_id", "default_value")
-    def onchange_copyvalue(self):
-        self.sub_object_id = False
-        self.copyvalue = False
-        if self.field_id and not self.field_id.relation:
-            self.copyvalue = "{{{{object.{} or {}}}}}".format(
-                self.field_id.name, self.default_value or "''"
-            )
-            self.sub_model_object_field_id = False
-        if self.field_id and self.field_id.relation:
-            self.sub_object_id = self.env["ir.model"].search(
-                [("model", "=", self.field_id.relation)]
-            )[0]
-        if self.sub_model_object_field_id:
-            self.copyvalue = "{{{{object.{}.{} or {}}}}}".format(
-                self.field_id.name,
-                self.sub_model_object_field_id.name,
-                self.default_value or "''",
-            )
 
     # compute the dynamic content for jinja expression
     def _compute_dynamic_content(self):
